// src/interfaces/auth.ts
// Generated by https://quicktype.io

export interface IUser {
    error:          boolean;
    mensajeRetorno: string;
    accessToken:    string;
    refreshToken:   string;
    user:           User;
}

export interface User {
    id:          number;
    name:        string;
    email:       string;
    password:    string;
    planId:      number;
    phoneNumber: string;
    role:        string;
    status:      string;
    isActive:    boolean;
    createdAt:   string;
    updatedAt:   string;
}